03-sep

+++ beecrowd/2300-2399/2355.c
@@ -1,10 +1,10 @@
 #include <stdio.h>
 int main()
 {
-    int n;
+    long long int n;
     while (1)
     {
-        scanf("%d", &n);
+        scanf("%lld", &n);
         if (n == 0)
         {
             break;
